refactor(revision): migrate to AsyncSession
